renal_project_final/web-project-main/server/scripts/requirements.txt
tensorflow>=2.10.0
opencv-python>=4.5.0
scipy>=1.7.0
numpy>=1.21.0
reportlab>=3.6.0
Pillow>=8.0.0 